# BE/openapi/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
...
paths:
...
  /tweets:
    post:
      tags:
        - tweets
      summary: Tạo tweet
      description: Tạo tweet
      operationId: createTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin tạo tweet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TweetRequestBody'
        required: true
      responses:
        default:
          description: Tạo tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tạo tweet thành công
    get:
      tags:
        - tweets
      summary: Lấy danh sách tweet từ những người mình theo dõi
      description: Lấy danh sách tweet từ những người mình theo dõi
      operationId: getTweets
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng tweet trên 1 trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        default:
          description: Lấy danh sách tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lấy danh sách tweet thành công
                  result:
                    type: object
                    properties:
                      tweets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tweet'
                      total_page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
...
  /medias/upload-image:
    post:
      tags:
        - medias
      summary: Upload ảnh
      description: Upload ảnh
      operationId: uploadImage
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload ảnh, upload 1 hoặc nhiều ảnh
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 4
        required: true
      responses:
        default:
          description: Upload ảnh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload ảnh thành công
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Media'
  /medias/upload-video:
    post:
      tags:
        - medias
      summary: Upload video
      description: Upload video
      operationId: uploadVideo
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload video, chỉ upload được 1 video
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
        required: true
      responses:
        default:
          description: Upload video thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload video thành công
                  result:
                    $ref: '#/components/schemas/Media'
  /medias/upload-video-hls:
    post:
      tags:
        - medias
      summary: Upload video HLS
      description: Upload video HLS
      operationId: uploadVideoHLS
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload video hls, chỉ upload được 1 video
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
        required: true
      responses:
        default:
          description: Upload video thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload video thành công
                  result:
                    $ref: '#/components/schemas/Media'
  /medias/video-status/{namevideo}:
    get:
      tags:
        - static
      summary: Lấy trạng thái upload video
      description: Lấy trạng thái upload video
      operationId: getVideoStatus
      parameters:
        - name: namevideo
          in: path
          description: Tên video
          required: true
          schema:
            type: string
            example: ZORfBrqSil1Ii1aNBHO_h
      security:
        - BearerAuth: []
      responses:
        default:
          description: Lấy trạng thái upload video thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStatus'
...